requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.3
clarifai==2.6.2
click==8.0.3
colorama==0.4.4
configparser==3.8.1
crayons==0.4.0
cryptography==35.0.0
decorator==4.4.2
Django==3.2.9
EasyProcess==0.3
emoji==1.4.1
explicit==0.1.3
Flask==2.0.2
future==0.18.2
googleapis-common-protos==1.53.0
grpcio==1.38.1
gunicorn==20.1.0
h11==0.12.0
huepy==1.2.1
idna==2.7
image==1.5.33
imageio==2.9.0
imageio-ffmpeg==0.4.5
iniconfig==1.1.1
instabot==0.117.0
instagramy==4.4
instaloader==4.8.1
instapy==0.6.13
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==2.6.0
MarkupSafe==2.0.1
MeaningCloud-python==2.0.0
mock==4.0.3
moviepy==1.0.3
numpy==1.21.1
outcome==1.1.0
packaging==21.0
pandas==1.3.0
pbr==5.6.0
pd==0.0.1
Pillow==8.4.0
pluggy==1.0.0
plyer==2.0.0
proglog==0.1.9
protobuf==3.17.3
py==1.10.0
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQRCode==1.2.1
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.18.0
python-telegram-bot==13.7
pytz==2021.1
PyVirtualDisplay==2.2
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
requests-toolbelt==0.8.0
responses==0.14.0
schedule==1.1.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
tk==0.1.0
toml==0.10.2
tornado==6.1
tqdm==4.61.2
trio==0.19.0
trio-websocket==0.9.2
tzlocal==2.1
urllib3==1.26.7
utils==1.0.1
webdriver-manager==3.5.0
webdriverdownloader==1.1.0.3
Werkzeug==2.0.2
wsproto==1.0.0
